# scripts/swagger.json
{
  "openapi": "3.0.1",
  "basePath": "/api",
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "info": {
    "title": "Document Management",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "header",
        "name": "Authorization",
        "description": "A JWT created using the same secret that is specified in the .env file."
      }
    },
...
    "parameters": {
      "docId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "The _id that matches the _id of a document in MongoDB.",
        "schema": {
          "type": "string"
        }
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
  "paths": {
...
    "/docs/{id}/{updateType}/{removeId}": {
      "delete": {
        "description": "Removes a project or jurisdiction id from those array properties, respectively",
        "parameters": [
          {
            "$ref": "#/components/parameters/docId"
          },
          {
            "name": "updateType",
            "in": "path",
            "required": true,
            "description": "The property in the Document from which to remove the :removeId parameter",
            "schema": {
              "type": "string",
              "enum": [
                "projects",
                "jurisdictions"
              ]
            }
          },
          {
            "name": "removeId",
            "in": "path",
            "required": true,
            "description": "The id to remove from the array property specified at :updateType",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns 200 on success"
          }
        }
      }
    },
    "/docs/{id}/{updateType}/{newId}": {
      "post": {
        "description": "Adds a new ID to the project or jurisdiction array for a document",
        "parameters": [
          {
            "$ref": "#/components/parameters/docId"
          },
          {
            "name": "updateType",
            "in": "path",
            "required": true,
            "description": "The property in the Document from which to add the :newId parameter",
            "schema": {
              "type": "string",
              "enum": [
                "projects",
                "jurisdictions"
              ]
            }
          },
          {
            "name": "newId",
            "in": "path",
            "required": true,
            "description": "The id to add to the the array property specified at :updateType",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns 200 on success"
          }
        }
      }
    },
...
  }
}